Split advice expression facts from transfers
